requirements.txt
pybit==5.8.0
python-dotenv==1.0.1
pandas==2.2.2
watchdog==4.0.0
prometheus_client==0.20.0
plotly==5.23.0
numpy==1.24.3
talib==0.4.0
requests==2.31.0
backtrader==1.9.76.123
matplotlib==3.7.1
setuptools==68.0.0